Renamed AuthUser.ts to auth-service.ts
